添加监控，增加snowflake uuid64位整形

+++ src/server/account/Account.go
@@ -6,7 +6,7 @@ import (
 
 type (
 	AccountDB struct{
-		AccountId int `sql:"primary;name:account_id"`//主键
+		AccountId int64 `sql:"primary;name:account_id"`//主键
 		AccountName string `sql:"name:account_name"`
 		Status int `sql:"name:status"`
 		LoginTime int64 `sql:"datetime;name:login_time""`//日期
